Moved classes to LootrunUtils
