--- common/src/main/java/com/wynntils/wc/utils/lootrun/objects/Point.java
@@ -1,9 +1,0 @@
-/*
- * Copyright © Wynntils 2022.
- * This file is released under AGPLv3. See LICENSE for full license details.
- */
-package com.wynntils.wc.utils.lootrun.objects;
-
-import net.minecraft.world.phys.Vec3;
-
-public record Point(Vec3 vec3, int color) {}
